pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "bayes-gp-llmops"
version = "0.1.0"
description = "Uncertainty-aware tiny LLaMA-style text classification with MLOps tooling."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "bayes-gp-llmops contributors" }
]
dependencies = [
  "torch>=2.4.0",
  "transformers>=4.45.0",
  "datasets>=2.21.0",
  "tokenizers>=0.20.0",
  "PyYAML>=6.0.2",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.5.0",
  "pydantic-settings>=2.5.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "mlflow>=2.16.0",
  "optuna>=4.0.0",
  "optuna-integration>=4.0.0",
  "matplotlib>=3.9.0",
  "httpx>=0.27.0",
  "streamlit>=1.44.0",
]

[project.scripts]
bayes-train = "bayes_gp_llmops.cli:run_train"
bayes-evaluate = "bayes_gp_llmops.cli:run_evaluate"
bayes-tune = "bayes_gp_llmops.cli:run_tune"
bayes-serve = "bayes_gp_llmops.cli:run_serve"
bayes-promote = "bayes_gp_llmops.cli:run_promote"
bayes-validate-bundle = "bayes_gp_llmops.cli:run_validate_bundle"
bayes-download-data = "bayes_gp_llmops.data.download:main"

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "pre-commit>=3.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bayes_gp_llmops"]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "scripts", "tests"]
exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "A", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
explicit_package_bases = true
mypy_path = "src"
files = ["src", "scripts", "tests"]

[[tool.mypy.overrides]]
module = [
  "uvicorn.*",
  "fastapi.*",
  "starlette.*",
  "datasets.*",
  "mlflow.*",
  "optuna.*",
  "optuna_integration.*",
  "sklearn.*",
  "tokenizers.*",
  "yaml.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
